Fix margin bug
LavaLauncher always creates surfaces spanning the entire width of the
outputs edge it is docked to. To prevent moving the surface outside the
boundaries of the output, only the margins not parallel to that length
are "real" layer shell surfaces. The margins parallel to that length are
simply "faked" by adjusting the x and y offsets of the draw code.