Improve margins
Margins in full mode now control the distance between the bar and the
screen edge instead of the item placement.

Opposite margins in default mode are no longer subtracted for alignments
other than center, meaning (for example) that having a left margin and
the exact same margin on the right side for a horizontal bar with end
alignment will result in space between the bar and the screen edge.