Workaround for limitations of layer shell keyboard interactivity
tl;dr: layer shell keyboard interactivity sucks ass for our use case.

We want to allow binds to use keyboard modifiers. To get keyboard
modifiers, we need to get the keyboard focus. This is where the "fun"
begins. If we request keyboard interactivity and are using the top or
overlay layer, we /always/ have the keyboard focus, meaning we have
to request interactivity on pointer enter and un-request it on pointer
leave. This will break like a cheap plastic toy accidentally dropped
into a blender as soon as multiple seats are involved and interact
with the bar simultaneously. Even more "fun", to get keyboard focus
on bottom and background layers, the bar has to be explicitly focused.
On most compositors this means clicking it. So just forget about all
your fancy scroll plus modifier binds.

The layer shell needs a replacement.