Deprecate "aggressive" mode and alignments.
LavaLauncher is not, was never intended to be and will never become a
generic widget thing. Let's get a bit more opinionted about this,
reducing code complexity, hopefully reducing maintenance effort and
making planned features easier to implement.

Bar alignment will be replaced with per-item alignment in a future
commit.